Renaming CurveImpl to CurveBackend and moving to package de.uni_kl.cs.discodnc.
