+++ src/main/java/de/uni_kl/cs/discodnc/CurveBackend.java
@@ -1,0 +1,48 @@
+/*
+ * This file is part of the Disco Deterministic Network Calculator.
+ *
+ * Copyright (C) 2013 - 2018 Steffen Bondorf
+ * Copyright (C) 2017+ The DiscoDNC contributors
+ *
+ * Distributed Computer Systems (DISCO) Lab
+ * University of Kaiserslautern, Germany
+ *
+ * http://discodnc.cs.uni-kl.de
+ *
+ *
+ * The Disco Deterministic Network Calculator (DiscoDNC) is free software;
+ * you can redistribute it and/or modify it under the terms of the 
+ * GNU Lesser General Public License as published by the Free Software Foundation; 
+ * either version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ */
+package de.uni_kl.cs.discodnc;
+
+import de.uni_kl.cs.discodnc.curves.CurvePwAffine;
+import de.uni_kl.cs.discodnc.curves.LinearSegment;
+import de.uni_kl.cs.discodnc.minplus.MinPlus;
+import de.uni_kl.cs.discodnc.numbers.Num;
+
+public interface CurveBackend {
+	
+	MinPlus getMinPlus();
+	
+	CurvePwAffine getCurve();
+	
+	LinearSegment createLinearSegment(Num x, Num y, Num grad, boolean leftopen);
+	
+	LinearSegment createHorizontalLine(double y);
+	
+	default void checkDependencies() {
+		
+	}
+}
