+++ src/test/java/de/uni_kl/cs/discodnc/DncTestArguments.java
@@ -30,7 +30,7 @@ package de.uni_kl.cs.discodnc;
 
 import de.uni_kl.cs.discodnc.nc.AnalysisConfig.ArrivalBoundMethod;
 import de.uni_kl.cs.discodnc.nc.CurveImpl_DNC;
-import de.uni_kl.cs.discodnc.Calculator.CurveImpl;
+import de.uni_kl.cs.discodnc.CurveBackend;
 import de.uni_kl.cs.discodnc.Calculator.NumImpl;
 //import CalculatorConfig.OperationImpl;
 import de.uni_kl.cs.discodnc.nc.AnalysisConfig;
@@ -61,7 +61,7 @@ public class DncTestArguments implements ArgumentsProvider {
 		nums.add(NumImpl.RATIONAL_INTEGER);
 		nums.add(NumImpl.RATIONAL_BIGINTEGER);
 
-		Set<CurveImpl> curves = new HashSet<CurveImpl>();
+		Set<CurveBackend> curves = new HashSet<CurveBackend>();
 		curves.add(CurveImpl_DNC.DNC);
 		
 //		Set<OperationImpl> operations = new HashSet<OperationImpl>();
@@ -113,7 +113,7 @@ public class DncTestArguments implements ArgumentsProvider {
 		// AB, remove duplicate ABs, tbrl opt convolution, tbrl opt deconvolution, mux,
 		// global mux def, number class to use, curve class to use, operations class to
 		// use
-		for (CurveImpl curve : curves) {
+		for (CurveBackend curve : curves) {
 			for (NumImpl num : nums) {
 				for (Set<ArrivalBoundMethod> single_ab : single_abs_allMux) {
 					for (AnalysisConfig.Multiplexing mux : mux_disciplines) {
